requirements.txt
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
geoalchemy2==0.15.1
shapely==2.0.4
fastapi==0.111.0
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
pydantic-settings==2.2.1
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
email-validator==2.1.1
google-generativeai==0.8.5
